requirements.txt
pycryptodome
pyaudio
tk
